ADD(garbage_collector) —> single file structure for garbage colector
